Added the security of viewing others notes, next step timestamp the image upload

+++ src/MapMyNotesApplication/MapMyNotesApplication/controllers/show_note.py
@@ -1,7 +1,8 @@
-from flask import Blueprint, render_template, redirect, url_for
+from flask import Blueprint, render_template, redirect, url_for, session
 
 from MapMyNotesApplication.models.date_time_helper import DateTimeHelper
 from MapMyNotesApplication.models.note import Note
+from MapMyNotesApplication.models.session_helper import SessionHelper
 
 shownote = Blueprint('shownote', __name__)
 
@@ -11,7 +12,10 @@ POST = "POST"
 
 @shownote.route("/show_note/<note_id>", methods=[GET])
 def show_note(note_id):
+    session_helper = SessionHelper(session)
     note = Note.query.get(note_id)
+    if note.user_id is not session_helper.return_user_id():
+        return "Sorry you can't access that"
     image_path = note.image_path
     module_code = note.meta_data.module_code.module_code
     lecturer = note.meta_data.lecturer
@@ -31,8 +35,11 @@ def show_note(note_id):
 
 @shownote.route("/delete_note/<note_id>", methods=[POST])
 def delete_note(note_id):
+    session_helper = SessionHelper(session)
     note = Note.query.get(note_id)
+    if note.user_id is not session_helper.return_user_id():
+        return "Sorry you cant access that"
     note.delete()
-    #TODO delete from calendar too
+    # TODO delete from calendar too
 
     return redirect(url_for("homepage.home_page_route"))
